+++ panther/cli/run_command.py
@@ -1,9 +1,8 @@
 import os
 
 import uvicorn
-from rich import print as rprint
 
-from panther.cli.utils import run_help_message
+from panther.cli.utils import print_uvicorn_help_message
 
 
 def _handle_commands(args: dict[str, str | None]) -> dict:
@@ -68,7 +67,7 @@ def _handle_commands(args: dict[str, str | None]) -> dict:
 
 def run(args: dict[str, str | None]) -> None:
     if any(a in args for a in ['h', 'help', '-h', '--help']):
-        rprint(run_help_message)
+        print_uvicorn_help_message()
         return
     command = {'app_dir': os.getcwd()}
     command.update(_handle_commands(args))